backend/requirements.txt
python-dotenv
pyyaml
pandas
numpy
scikit-learn
pyarrow
pydantic==2.4.2
pydantic[email]==2.4.2
requests
httpx==0.25.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.1
structlog==23.3.0
psycopg2-binary==2.9.9
pycountry==23.12.11
python-multipart==0.0.6
aiosqlite==0.22.1
